# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ActionPath API – personalized recommendations for Indian government schemes, documents, and jobs.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ActionPriority =
  (typeof ActionPriority)[keyof typeof ActionPriority];

export const ActionPriority = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Action {
  id: number;
  category: string;
  categoryLabel: string;
  title: string;
  description: string;
  documents: string[];
  link: string;
  priority: ActionPriority;
  tags: string[];
  /** Human-readable eligibility summary. */
  eligibility?: string;
  /** Short summary of the benefit (e.g. "Up to ₹6,000/year"). */
  benefit?: string;
  /** Issuing ministry / department. */
  ministry?: string;
}

export interface ActionListResponse {
  items: Action[];
  total: number;
}

export interface CategorySummary {
  slug: string;
  label: string;
  count: number;
}

export interface StatsResponse {
  totalActions: number;
  totalCategories: number;
  highPriorityCount: number;
}

export type UserProfileGender =
  (typeof UserProfileGender)[keyof typeof UserProfileGender];

export const UserProfileGender = {
  male: "male",
  female: "female",
  other: "other",
} as const;

export type UserProfileOccupation =
  (typeof UserProfileOccupation)[keyof typeof UserProfileOccupation];

export const UserProfileOccupation = {
  student: "student",
  salaried: "salaried",
  self_employed: "self_employed",
  farmer: "farmer",
  unemployed: "unemployed",
  homemaker: "homemaker",
  retired: "retired",
  daily_wage: "daily_wage",
} as const;

export type UserProfileEducation =
  (typeof UserProfileEducation)[keyof typeof UserProfileEducation];

export const UserProfileEducation = {
  none: "none",
  primary: "primary",
  secondary: "secondary",
  higher_secondary: "higher_secondary",
  graduate: "graduate",
  postgraduate: "postgraduate",
} as const;

export type UserProfileMaritalStatus =
  (typeof UserProfileMaritalStatus)[keyof typeof UserProfileMaritalStatus];

export const UserProfileMaritalStatus = {
  single: "single",
  married: "married",
  divorced: "divorced",
  widowed: "widowed",
} as const;

export type UserProfileHasAadhaar =
  (typeof UserProfileHasAadhaar)[keyof typeof UserProfileHasAadhaar];

export const UserProfileHasAadhaar = {
  yes: "yes",
  no: "no",
} as const;

export type UserProfileHasPAN =
  (typeof UserProfileHasPAN)[keyof typeof UserProfileHasPAN];

export const UserProfileHasPAN = {
  yes: "yes",
  no: "no",
} as const;

export type UserProfileHasVoterId =
  (typeof UserProfileHasVoterId)[keyof typeof UserProfileHasVoterId];

export const UserProfileHasVoterId = {
  yes: "yes",
  no: "no",
} as const;

export type UserProfileHasRationCard =
  (typeof UserProfileHasRationCard)[keyof typeof UserProfileHasRationCard];

export const UserProfileHasRationCard = {
  yes: "yes",
  no: "no",
} as const;

export type UserProfileHasDrivingLicense =
  (typeof UserProfileHasDrivingLicense)[keyof typeof UserProfileHasDrivingLicense];

export const UserProfileHasDrivingLicense = {
  yes: "yes",
  no: "no",
} as const;

export type UserProfileHasBankAccount =
  (typeof UserProfileHasBankAccount)[keyof typeof UserProfileHasBankAccount];

export const UserProfileHasBankAccount = {
  yes: "yes",
  no: "no",
} as const;

export interface UserProfile {
  name?: string;
  /**
   * @minimum 0
   * @maximum 120
   */
  age: number;
  gender: UserProfileGender;
  occupation: UserProfileOccupation;
  /** Annual household income in INR. */
  income: number;
  state: string;
  education: UserProfileEducation;
  maritalStatus: UserProfileMaritalStatus;
  differentlyAbled?: boolean;
  hasAadhaar: UserProfileHasAadhaar;
  hasPAN: UserProfileHasPAN;
  hasVoterId: UserProfileHasVoterId;
  hasRationCard: UserProfileHasRationCard;
  hasDrivingLicense: UserProfileHasDrivingLicense;
  hasBankAccount: UserProfileHasBankAccount;
}

export type ListActionsParams = {
  /**
   * Filter by category slug (e.g. "documents", "schemes", "jobs", "women", "farmer", "student", "youth", "health", "housing", "senior", "finance").
   */
  category?: string;
  /**
   * Free-text search over title, description, and tags.
   */
  search?: string;
};