# lib/api-zod/src/generated/types/action.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ActionPath API – personalized recommendations for Indian government schemes, documents, and jobs.
 * OpenAPI spec version: 0.1.0
 */
import type { ActionPriority } from "./actionPriority";

export interface Action {
  id: number;
  category: string;
  categoryLabel: string;
  title: string;
  description: string;
  documents: string[];
  link: string;
  priority: ActionPriority;
  tags: string[];
  /** Human-readable eligibility summary. */
  eligibility?: string;
  /** Short summary of the benefit (e.g. "Up to ₹6,000/year"). */
  benefit?: string;
  /** Issuing ministry / department. */
  ministry?: string;
}