# lib/api-zod/src/generated/types/actionListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ActionPath API – personalized recommendations for Indian government schemes, documents, and jobs.
 * OpenAPI spec version: 0.1.0
 */
import type { Action } from "./action";

export interface ActionListResponse {
  items: Action[];
  total: number;
}