# lib/api-zod/src/generated/types/actionPriority.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ActionPath API – personalized recommendations for Indian government schemes, documents, and jobs.
 * OpenAPI spec version: 0.1.0
 */

export type ActionPriority =
  (typeof ActionPriority)[keyof typeof ActionPriority];

export const ActionPriority = {
  high: "high",
  medium: "medium",
  low: "low",
} as const;