# lib/api-zod/src/generated/types/categorySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ActionPath API – personalized recommendations for Indian government schemes, documents, and jobs.
 * OpenAPI spec version: 0.1.0
 */

export interface CategorySummary {
  slug: string;
  label: string;
  count: number;
}