# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ActionPath API – personalized recommendations for Indian government schemes, documents, and jobs.
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileEducation } from "./userProfileEducation";
import type { UserProfileGender } from "./userProfileGender";
import type { UserProfileHasAadhaar } from "./userProfileHasAadhaar";
import type { UserProfileHasBankAccount } from "./userProfileHasBankAccount";
import type { UserProfileHasDrivingLicense } from "./userProfileHasDrivingLicense";
import type { UserProfileHasPAN } from "./userProfileHasPAN";
import type { UserProfileHasRationCard } from "./userProfileHasRationCard";
import type { UserProfileHasVoterId } from "./userProfileHasVoterId";
import type { UserProfileMaritalStatus } from "./userProfileMaritalStatus";
import type { UserProfileOccupation } from "./userProfileOccupation";

export interface UserProfile {
  name?: string;
  /**
   * @minimum 0
   * @maximum 120
   */
  age: number;
  gender: UserProfileGender;
  occupation: UserProfileOccupation;
  /** Annual household income in INR. */
  income: number;
  state: string;
  education: UserProfileEducation;
  maritalStatus: UserProfileMaritalStatus;
  differentlyAbled?: boolean;
  hasAadhaar: UserProfileHasAadhaar;
  hasPAN: UserProfileHasPAN;
  hasVoterId: UserProfileHasVoterId;
  hasRationCard: UserProfileHasRationCard;
  hasDrivingLicense: UserProfileHasDrivingLicense;
  hasBankAccount: UserProfileHasBankAccount;
}